pyproject.toml
[tool.poetry]
name = "medical-chatbot-ai"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.10"
langgraph = "^0.2.48"
langchain-openai = "^0.2.8"
langchain-community = "^0.3.7"
sentence-transformers = "^3.3.0"
einops = "^0.8.0"
pypdf = "^5.1.0"
chromadb = "^0.5.18"
langchain-groq = "^0.2.1"
ipython = "^8.29.0"
fpdf = "^1.7.2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

